backend/requirements/dev.txt
-r base.txt

# Development tools
django-debug-toolbar>=4.2,<5.0
django-extensions>=3.2,<4.0
ipython>=8.18,<9.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=21.0,<22.0

# Code Quality
black>=23.12,<24.0
isort>=5.13,<6.0
flake8>=6.1,<7.0
mypy>=1.7,<2.0
django-stubs>=4.2,<5.0
djangorestframework-stubs>=3.14,<4.0

# Pre-commit
pre-commit>=3.6,<4.0
